level 1 and 2 and tutorials are done.  stop touching them!

+++ Source/Game/Components/EnemyComponent.swift
@@ -7,7 +7,7 @@
 //
 
 class EnemyComponent: Component {
-    var experience: Float = 0
+    var experience: Int = 0
 
     typealias OnAttacked = (projectile: Node) -> Void
     var _onAttacked: [OnAttacked] = []
